Refactor iteration callback implementation to follow code base standards
